fix: extra protection against untrusted input
Although the input variables come from the workflow files, and thus may
be trusted, it's also possible for workflows to be poorly secured.
Avoiding use of ${{ }} substitutions in this action can provide some
protection in those cases by placing inputs in environment variables,
and using those instead of the more dangerous substitions.

On the other hand, it dangerous to place non-default GitHub token
values in the environment; placing it only in the environment of the
GitHub CLI command rather than the general environment visible to all
workflow components reduces that risk.

Using input substitution for the token carries some risk, but that risk
can be mitigated more easily than protecting against exfiltration of a
GitHub token with write permission and access to secrets.

https://securitylab.github.com/research/github-actions-untrusted-input/
has more details about these security issues.